Fix a bug involving the reuse of invalid data.
The `split_odd` function wasn't zeroing the second half of the byte where the split
was being performed. This meant that later calls to methods like `push` - which assume
invalid bytes to be 0 - ended up mixing in stale values (via binary OR).

This patch double fixes the bug by zeroing in both locations. There don't seem to be any
other problematic versions of this problem, but this has shaken my faith in the correctness
of this library. A careful audit would be nice.